requirements.txt
alembic==1.13.1
asyncpg==0.29.0
databases==0.9.0
fastapi==0.111.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.29
uvicorn==0.29.0